Refactoring file ops.
